+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Use fully-qualified resource name for generating manifests, to avoid conflicts (https://github.com/pulumi/pulumi-kubernetes/pull/2007)
+- Upgrade helm and k8s client-go module dependencies (https://github.com/pulumi/pulumi-kubernetes/pull/2008)
 
 ## 3.19.3 (June 8, 2022)